orginize destination tests
